# daemon/algod/api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API Endpoint for AlgoD Operations.",
    "title": "Algod REST API.",
    "contact": {
      "email": "contact@algorand.com"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/",
  "paths": {
...
    "/v1/account/{address}/transactions": {
      "get": {
        "description": "Returns the list of confirmed transactions between within a date range. When indexer is disabled this call requires firstRound and lastRound and returns an error if firstRound is not available to the node. The transaction results start from the oldest round.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "summary": "Get a list of confirmed transactions.",
        "operationId": "Transactions",
        "parameters": [
          {
            "pattern": "[A-Z0-9]{58}",
            "type": "string",
            "description": "An account public key",
            "name": "address",
            "in": "path",
            "required": true
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int64",
            "description": "Do not fetch any transactions before this round.",
            "name": "firstRound",
            "in": "query"
          },
          {
            "minimum": 0,
            "type": "integer",
            "format": "int64",
            "description": "Do not fetch any transactions after this round.",
            "name": "lastRound",
            "in": "query"
          },
          {
            "type": "string",
            "format": "date",
            "description": "Do not fetch any transactions before this date. (enabled only with indexer)",
            "name": "fromDate",
            "in": "query"
          },
          {
            "type": "string",
            "format": "date",
            "description": "Do not fetch any transactions after this date. (enabled only with indexer)",
            "name": "toDate",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "description": "maximum transactions to show (default to 100)",
            "name": "max",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TransactionsResponse"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Invalid API Token"
          },
          "500": {
            "description": "Internal Error",
            "schema": {
              "type": "string"
            }
          },
          "default": {
            "description": "Unknown Error"
          }
        }
      }
    },
...
    "/v1/transaction/{txid}": {
      "get": {
        "description": "Returns the transaction information of the given txid. Works only if the indexer is enabled.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "summary": "Get an information of a single transaction.",
        "operationId": "Transaction",
        "parameters": [
          {
            "pattern": "[A-Z0-9]+",
            "type": "string",
            "description": "A transaction id",
            "name": "txid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TransactionResponse"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Invalid API Token"
          },
          "404": {
            "description": "Transaction Not Found",
            "schema": {
              "type": "string"
            }
          },
          "default": {
            "description": "Unknown Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApplicationCallTransactionType": {
      "description": "ApplicationCallTransactionType contains the additional fields for an ApplicationCall transaction",
      "type": "object",
      "required": [
        "id",
        "accounts",
        "foreignapps",
        "foreignassets",
        "appargs",
        "approvprog",
        "clearprog"
      ],
      "properties": {
        "accounts": {
          "description": "Accounts lists the accounts (in addition to the sender) that may be\naccessed from the application's ApprovalProgram and ClearStateProgram.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Accounts"
        },
        "appargs": {
          "description": "ApplicationArgs lists some transaction-specific arguments accessible\nfrom application logic",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "ApplicationArgs"
        },
        "approvprog": {
          "description": "ApprovalProgram determines whether or not this ApplicationCall\ntransaction will be approved or not. It does not execute when\nOnCompletion == ClearStateOC, because clearing local state is always\nallowed.",
          "type": "string",
          "x-go-name": "ApprovalProgram"
        },
        "clearprog": {
          "description": "ClearStateProgram executes when an ApplicationCall transaction\nexecutes with OnCompletion == ClearStateOC. However, this program\nmay not reject the transaction (only update state). If this program",
          "type": "string",
          "x-go-name": "ClearStateProgram"
        },
        "foreignapps": {
          "description": "ForeignApps lists the applications (in addition to txn.ApplicationID)\nwhose global states may be accessed by this application's\nApprovalProgram and ClearStateProgram. The access is read-only.",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint64"
          },
          "x-go-name": "ForeignApps"
        },
        "foreignassets": {
          "description": "ForeignAssets lists the assets whose parameters may be accessed by\nthis application's ApprovalProgram and ClearStateProgram. The access\nis read-only.",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint64"
          },
          "x-go-name": "ForeignAssets"
        },
        "globalschema": {
          "$ref": "#/definitions/StateSchema"
        },
        "id": {
          "description": "ApplicationID is the application being interacted with, or 0 if\ncreating a new application.",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ApplicationID"
        },
        "localschema": {
          "$ref": "#/definitions/StateSchema"
        },
        "oncompletion": {
          "description": "OnCompletion specifies what side effects this transaction will have\nif it successfully makes it into a block.\n\nrequire: true",
          "type": "string",
          "x-go-name": "OnCompletion"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "AssetConfigTransactionType": {
      "description": "AssetConfigTransactionType contains the additional fields for an asset config transaction",
      "type": "object",
      "properties": {
        "id": {
          "description": "AssetID is the asset being configured (or empty if creating)",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "AssetID"
        },
        "params": {
          "$ref": "#/definitions/AssetParams"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
    "AssetFreezeTransactionType": {
      "description": "AssetFreezeTransactionType contains the additional fields for an asset freeze transaction",
      "type": "object",
      "required": [
        "id",
        "acct",
        "freeze"
      ],
      "properties": {
        "acct": {
          "description": "Account specifies the account where the asset is being frozen or thawed.",
          "type": "string",
          "x-go-name": "Account"
        },
        "freeze": {
          "description": "NewFreezeStatus specifies the new freeze status.",
          "type": "boolean",
          "x-go-name": "NewFreezeStatus"
        },
        "id": {
          "description": "AssetID is the asset being configured (or empty if creating)",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "AssetID"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "AssetParams": {
      "type": "object",
      "title": "AssetParams specifies the parameters for an asset.",
      "required": [
        "creator",
        "total",
        "decimals"
      ],
      "properties": {
        "assetname": {
          "description": "AssetName specifies the name of this asset,\nas supplied by the creator.",
          "type": "string",
          "x-go-name": "AssetName"
        },
        "clawbackaddr": {
          "description": "ClawbackAddr specifies the address used to clawback holdings of\nthis asset.  If empty, clawback is not permitted.",
          "type": "string",
          "x-go-name": "ClawbackAddr"
        },
        "creator": {
          "description": "Creator specifies the address that created this asset.\nThis is the address where the parameters for this asset\ncan be found, and also the address where unwanted asset\nunits can be sent in the worst case.",
          "type": "string",
          "x-go-name": "Creator"
        },
        "decimals": {
          "description": "Decimals specifies the number of digits to use after the decimal\npoint when displaying this asset. If 0, the asset is not divisible.\nIf 1, the base unit of the asset is in tenths. If 2, the base unit\nof the asset is in hundredths, and so on.",
          "type": "integer",
          "format": "uint32",
          "x-go-name": "Decimals"
        },
        "defaultfrozen": {
          "description": "DefaultFrozen specifies whether holdings in this asset\nare frozen by default.",
          "type": "boolean",
          "x-go-name": "DefaultFrozen"
        },
        "freezeaddr": {
          "description": "FreezeAddr specifies the address used to freeze holdings of\nthis asset.  If empty, freezing is not permitted.",
          "type": "string",
          "x-go-name": "FreezeAddr"
        },
        "managerkey": {
          "description": "ManagerAddr specifies the address used to manage the keys of this\nasset and to destroy it.",
          "type": "string",
          "x-go-name": "ManagerAddr"
        },
        "metadatahash": {
          "description": "MetadataHash specifies a commitment to some unspecified asset\nmetadata. The format of this metadata is up to the application.",
          "type": "string",
          "format": "byte",
          "x-go-name": "MetadataHash"
        },
        "reserveaddr": {
          "description": "ReserveAddr specifies the address holding reserve (non-minted)\nunits of this asset.",
          "type": "string",
          "x-go-name": "ReserveAddr"
        },
        "total": {
          "description": "Total specifies the total number of units of this asset.",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Total"
        },
        "unitname": {
          "description": "UnitName specifies the name of a unit of this asset,\nas supplied by the creator.",
          "type": "string",
          "x-go-name": "UnitName"
        },
        "url": {
          "description": "URL specifies a URL where more information about the asset can be\nretrieved",
          "type": "string",
          "x-go-name": "URL"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
    "AssetTransferTransactionType": {
      "description": "AssetTransferTransactionType contains the additional fields for an asset transfer transaction",
      "type": "object",
      "required": [
        "id",
        "amt",
        "rcv"
      ],
      "properties": {
        "amt": {
          "description": "Amount is the amount being transferred.",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Amount"
        },
        "closeto": {
          "description": "CloseTo is the destination for remaining funds (if closing).",
          "type": "string",
          "x-go-name": "CloseTo"
        },
        "closetoamount": {
          "description": "CloseToAmount is amount of the remaining funds that were transferred to the close to address (if closing).",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CloseToAmount"
        },
        "id": {
          "description": "AssetID is the asset being configured (or empty if creating)",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "AssetID"
        },
        "rcv": {
          "description": "Receiver is the recipient account.",
          "type": "string",
          "x-go-name": "Receiver"
        },
        "snd": {
          "description": "Sender is the source account (if using clawback).",
          "type": "string",
          "x-go-name": "Sender"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "CompactCertTransactionType": {
      "description": "CompactCertTransactionType contains the additional fields for a compact cert transaction",
      "type": "object",
      "required": [
        "rnd",
        "cert"
      ],
      "properties": {
        "cert": {
          "description": "Cert is the msgpack encoding of the compact cert.",
          "type": "string",
          "format": "byte",
          "x-go-name": "Cert"
        },
        "rnd": {
          "description": "CertRound is the round whose block this compact cert refers to.",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CertRound"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
    "KeyregTransactionType": {
      "description": "KeyregTransactionType contains the additional fields for a keyreg Transaction",
      "type": "object",
      "properties": {
        "selkey": {
          "description": "SelectionPK is the VRF public key used in key registration transactions",
          "type": "string",
          "format": "byte",
          "x-go-name": "SelectionPK"
        },
        "votefst": {
          "description": "VoteFirst is the first round this participation key is valid",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "VoteFirst"
        },
        "votekd": {
          "description": "VoteKeyDilution is the dilution for the 2-level participation key",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "VoteKeyDilution"
        },
        "votekey": {
          "description": "VotePK is the participation public key used in key registration transactions",
          "type": "string",
          "format": "byte",
          "x-go-name": "VotePK"
        },
        "votelst": {
          "description": "VoteLast is the last round this participation key is valid",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "VoteLast"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "PaymentTransactionType": {
      "description": "PaymentTransactionType contains the additional fields for a payment Transaction",
      "type": "object",
      "required": [
        "to",
        "amount"
      ],
      "properties": {
        "amount": {
          "description": "Amount is the amount of MicroAlgos intended to be transferred",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Amount"
        },
        "close": {
          "description": "CloseRemainderTo is the address the sender closed to",
          "type": "string",
          "x-go-name": "CloseRemainderTo"
        },
        "closeamount": {
          "description": "CloseAmount is the amount sent to CloseRemainderTo, for committed transaction",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CloseAmount"
        },
        "closerewards": {
          "description": "CloseRewards is the amount of pending rewards applied to the CloseRemainderTo\naccount as part of this transaction.",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CloseRewards"
        },
        "to": {
          "description": "To is the receiver's address",
          "type": "string",
          "x-go-name": "To"
        },
        "torewards": {
          "description": "ToRewards is the amount of pending rewards applied to the To account\nas part of this transaction.",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ToRewards"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "StateSchema": {
      "description": "swagger: model StateSchema",
      "type": "object",
      "title": "StateSchema represents a LocalStateSchema or GlobalStateSchema. These\nschemas determine how much storage may be used in a LocalState or\nGlobalState for an application. The more space used, the larger minimum\nbalance must be maintained in the account holding the data.",
      "required": [
        "uints",
        "byteslices"
      ],
      "properties": {
        "byteslices": {
          "description": "NumByteSlice is the maximum number of TEAL byte slices that may be\nstored in the key/value store",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "NumByteSlice"
        },
        "uints": {
          "description": "NumUint is the maximum number of TEAL uints that may be stored in\nthe key/value store",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "NumUint"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "Transaction": {
      "description": "Transaction contains all fields common to all transactions and serves as an envelope to all transactions\ntype",
      "type": "object",
      "required": [
        "type",
        "tx",
        "from",
        "fee",
        "first-round",
        "last-round",
        "app",
        "compactcert",
        "genesisID",
        "genesishashb64"
      ],
      "properties": {
        "app": {
          "$ref": "#/definitions/ApplicationCallTransactionType"
        },
        "compactcert": {
          "$ref": "#/definitions/CompactCertTransactionType"
        },
        "curcfg": {
          "$ref": "#/definitions/AssetConfigTransactionType"
        },
        "curfrz": {
          "$ref": "#/definitions/AssetFreezeTransactionType"
        },
        "curxfer": {
          "$ref": "#/definitions/AssetTransferTransactionType"
        },
        "fee": {
          "description": "Fee is the transaction fee",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Fee"
        },
        "first-round": {
          "description": "FirstRound indicates the first valid round for this transaction",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "FirstRound"
        },
        "from": {
          "description": "From is the sender's address",
          "type": "string",
          "x-go-name": "From"
        },
        "fromrewards": {
          "description": "FromRewards is the amount of pending rewards applied to the From\naccount as part of this transaction.",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "FromRewards"
        },
        "genesisID": {
          "description": "Genesis ID",
          "type": "string",
          "x-go-name": "GenesisID"
        },
        "genesishashb64": {
          "description": "Genesis hash",
          "type": "string",
          "format": "byte",
          "x-go-name": "GenesisHash"
        },
        "group": {
          "description": "Group",
          "type": "string",
          "format": "byte",
          "x-go-name": "Group"
        },
        "keyreg": {
          "$ref": "#/definitions/KeyregTransactionType"
        },
        "last-round": {
          "description": "LastRound indicates the last valid round for this transaction",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "LastRound"
        },
        "lease": {
          "description": "Lease enforces mutual exclusion of transactions.  If this field is\nnonzero, then once the transaction is confirmed, it acquires the\nlease identified by the (Sender, Lease) pair of the transaction until\nthe LastValid round passes.  While this transaction possesses the\nlease, no other transaction specifying this lease can be confirmed.",
          "type": "string",
          "format": "byte",
          "x-go-name": "Lease"
        },
        "noteb64": {
          "description": "Note is a free form data",
          "type": "string",
          "format": "byte",
          "x-go-name": "Note"
        },
        "payment": {
          "$ref": "#/definitions/PaymentTransactionType"
        },
        "poolerror": {
          "description": "PoolError indicates the transaction was evicted from this node's transaction\npool (if non-empty).  A non-empty PoolError does not guarantee that the\ntransaction will never be committed; other nodes may not have evicted the\ntransaction and may attempt to commit it in the future.",
          "type": "string",
          "x-go-name": "PoolError"
        },
        "round": {
          "description": "ConfirmedRound indicates the block number this transaction appeared in",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ConfirmedRound"
        },
        "tx": {
          "description": "TxID is the transaction ID",
          "type": "string",
          "x-go-name": "TxID"
        },
        "txresults": {
          "$ref": "#/definitions/TransactionResults"
        },
        "type": {
          "description": "Type is the transaction type",
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "TransactionList": {
      "description": "TransactionList contains a list of transactions",
      "type": "object",
      "required": [
        "transactions"
      ],
      "properties": {
        "transactions": {
          "description": "TransactionList is a list of transactions",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Transaction"
          },
          "x-go-name": "Transactions"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
    "TransactionResults": {
      "description": "TransactionResults contains information about the side effects of a transaction",
      "type": "object",
      "properties": {
        "createdapp": {
          "description": "CreatedAppIndex indicates the app index of an app created by this txn",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CreatedAppIndex"
        },
        "createdasset": {
          "description": "CreatedAssetIndex indicates the asset index of an asset created by this txn",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CreatedAssetIndex"
        }
      },
      "x-go-package": "github.com/Orca18/novarand/daemon/algod/api/spec/v1"
    },
...
  },
  "responses": {
...
    "TransactionResponse": {
      "description": "TransactionResponse contains a transaction information",
      "schema": {
        "$ref": "#/definitions/Transaction"
      }
    },
    "TransactionsResponse": {
      "description": "TransactionsResponse contains a list of transactions",
      "schema": {
        "$ref": "#/definitions/TransactionList"
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "description": "Generated header parameter. This token can be generated using the Goal command line tool. Example value ='b7e384d0317b8050ce45900a94a1931e28540e1f69b2d242b424659c341b4697'",
      "type": "apiKey",
      "name": "X-Algo-API-Token",
      "in": "header",
      "x-example": "b7e384d0317b8050ce45900a94a1931e28540e1f69b2d242b424659c341b4697"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}